tts_server/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
chatterbox-tts
torch
torchaudio
